distributed-messaging-queue-rabbitmq/requirements.txt
# RabbitMQ client library
pika==1.3.2

# Configuration management
python-dotenv==1.0.0

# Testing libraries
pytest==7.4.0
pytest-cov==4.1.0
mock==5.1.0

# Utility libraries
requests==2.31.0
